Consolidate filter bubbles into a single AppliedFilters component
It turned out that the ResultContainer was only providing props used for rendering applied facets, so I just renamed it to AppliedFacetsContainer.

Also we trim 'T00:00:00Z' off of applied time filters now.
